requirements.txt
Django==1.6.6
Fabric==1.9.1
South==1.0
Unidecode==0.04.16
django-crispy-forms==1.4.0
django-sendfile==0.3.6
ecdsa==0.11
paramiko==1.14.0
pycrypto==2.6.1
wsgiref==0.1.2
